+++ examples/rainbowkit-tailwind/pages/[chainId]/[contractAddress]/index.tsx
@@ -1,0 +1,13 @@
+import { useRouter } from "next/router";
+import { useState } from "react"
+import ContractPage from "../../../components/ContractPage"
+
+const Page = () => {
+    const router = useRouter()
+    const {contractAddress, chainId} = router.query
+    const [metadata, setMetadata] = useState({})
+
+    return <ContractPage contractAddress={contractAddress} chainId={chainId} metadata={metadata} />
+}
+
+export default Page