+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.3] - 2022-11-26
+
+### evaluate function
+
+- composite of all evaluators.
+- provides an array of scores.
+- when a new evaluator is added, it should be called in the `evaluate` function.
+
 ## [0.0.2] - 2022-11-25
 
 ### functional mvp